unified GLTFComponent as Async to test WebGL
