+++ unity-renderer/Assets/DCLPlugins/BuilderInWorld/Tests/BIWCreatorShould.cs
@@ -3,14 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using DCL;
 using DCL.Builder;
-using DCL.Components;
 using DCL.Controllers;
 using DCL.Helpers;
 using DCL.Models;
+using MainScripts.DCL.GLTF;
 using NUnit.Framework;
 using UnityEngine;
-using UnityEngine.TestTools;
-using UnityGLTF;
 
 public class BIWCreatorShould : IntegrationTestSuite_Legacy
 {
