+++ unity-renderer/Assets/DCLPlugins/BuilderInWorld/Tests/BIWFloorHandlerShould.cs
@@ -1,16 +1,13 @@
 ﻿using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
 using DCL;
 using DCL.Builder;
 using DCL.Components;
 using DCL.Controllers;
 using DCL.Helpers;
 using DCL.Models;
+using MainScripts.DCL.GLTF;
 using NUnit.Framework;
 using UnityEngine;
-using UnityEngine.TestTools;
-using UnityGLTF;
 
 public class BIWFloorHandlerShould : IntegrationTestSuite_Legacy
 {
