+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Components/LoadableShapes/Tests/GLTFImporterVisualTests.cs
@@ -2,6 +2,7 @@
 using DCL.Components;
 using DCL.Helpers;
 using DCL.Models;
+using MainScripts.DCL.GLTF;
 using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
