+++ unity-renderer/Assets/UnityGLTF/Scripts/Editor/UnityGLTFEditorAssembly.asmdef
@@ -1,7 +1,9 @@
 {
     "name": "UnityGLTFEditorAssembly",
+    "rootNamespace": "",
     "references": [
-        "GUID:54b09a4a10c111e46a817873fcea53fe"
+        "GUID:54b09a4a10c111e46a817873fcea53fe",
+        "GUID:f51ebe6a0ceec4240a699833d6309b23"
     ],
     "includePlatforms": [
         "Editor"
