--- unity-renderer/Assets/UnityGLTF/Scripts/GLTFComponentAsync.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 2f4ff1679c90446f98015a2f4270db10
-timeCreated: 1641313276
