--- unity-renderer/Assets/UnityGLTF/Scripts/GLTFSceneImporterAsync.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 2d3cf980a62347e89c00c29b83ccfcd5
-timeCreated: 1641326249
