--- unity-renderer/Assets/UnityGLTF/Scripts/Loader/ILoaderAsync.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 525f3459e1ecb45ebbd38b70fe758ae6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
