--- unity-renderer/Assets/UnityGLTF/Scripts/Loader/WebRequestLoaderAsync.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5ba74a13857c5474da5f9d6394af4813
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
