+++ unity-renderer/Assets/UnityGLTF/Tests/GLTFBenchmark.cs
@@ -1,5 +1,6 @@
 using DCL.Helpers;
 using System.Collections;
+using MainScripts.DCL.GLTF;
 using UnityEngine;
 using UnityGLTF;
